Add setup for blog header
