+++ client/src/layouts/Sidebar/BlogNav.tsx
@@ -4,7 +4,7 @@ export function BlogNav() {
       <svg viewBox="0 -9 3 24" className="overflow-visible mr-3 text-slate-400 w-auto h-6 group-hover:text-slate-600 dark:group-hover:text-slate-300">
         <path d="M3 0L0 3L3 6" fill="none" stroke="currentColor" stroke-width="2" stroke-linecap="round" stroke-linejoin="round"></path>
       </svg>
-      Go back
+      See more posts
     </a>
   </>
 }
