+++ client/src/ui/MDXContentController/PageHeader.tsx
@@ -1,14 +1,20 @@
+import { useCurrentPath } from '@/hooks/useCurrentPath';
+import { Meta } from '@/types/MDXContentControllerTypes';
 import { UserFeedback } from '@/ui/Feedback';
+import { slugToTitle } from '@/utils/titleText/slugToTitle';
 
 type PageHeaderProps = {
-  title?: string;
-  description?: string;
   section: string;
-  isBlogMode?: boolean;
+  meta: Meta;
 };
 
-export function PageHeader({ title, description, section, isBlogMode }: PageHeaderProps) {
+export function PageHeader({ section, meta }: PageHeaderProps) {
+  const currentPath = useCurrentPath();
+  const title = meta.title || slugToTitle(currentPath)
+  const description = meta.description
   if (!title && !description) return null;
+
+  const isBlogMode = meta.mode === 'blog';
 
   return (
     <header id="header" className="relative z-20">
@@ -34,15 +40,19 @@ export function PageHeader({ title, description, section, isBlogMode }: PageHead
       {description && (
         <p className="mt-2 text-lg text-slate-700 dark:text-slate-400">{description}</p>
       )}
-      <li className="flex items-center font-medium whitespace-nowrap mt-6">
-        <img src="https://mintlify.com/_next/image?url=%2F_next%2Fstatic%2Fmedia%2Fdeclan.70da18ee.jpeg&w=96&q=75" alt="" className="mr-3 w-9 h-9 rounded-full bg-slate-50 dark:bg-slate-800 border border-slate-600 dark:border-slate-400" decoding="async" />
+      {
+        isBlogMode && <div className="flex items-center font-medium whitespace-nowrap mt-6">
+        <img src="https://mintlify.com/_next/image?url=%2F_next%2Fstatic%2Fmedia%2Fdeclan.70da18ee.jpeg&w=96&q=75" alt="" className="mr-3 w-9 h-9 rounded-full bg-slate-50 dark:bg-slate-800 border border-slate-400 dark:border-slate-500" decoding="async" />
           <div className="text-sm leading-4">
             <div className="text-slate-900 dark:text-slate-200">Han Wang</div>
             <div className="mt-1">
-              <a href="https://twitter.com/adamwathan" className="text-primary hover:text-primary-dark dark:text-primary-light">3 min read</a>
+              <div className="flex items-center text-primary dark:text-primary-light">
+                August 30 <div className="mx-1 text-[0.5rem] text-slate-400 dark:text-slate-500">•</div> 3 min read
+              </div>
             </div>
           </div>
-        </li>
+        </div>
+      }
     </header>
   );
 }
